add some functions to utils
